clean slate and legendary spirit fix
Clean slate scrolls could not be used on equips with 0 slots remaining. Additionally, using a scroll with legendary spirit on an equip with 0 slots remaining would freeze the legendary spirit window. This fixes both issues by adding an isCleanSlate check and sending a different packet in the legendary spirit case (the player will still need to dispose after).